Rename features/hololive --> features/group
