+++ web/features/group/chart/api/getChartOfChannels.ts
