+++ web/features/group/schedule/components/Schedule.tsx
@@ -1,6 +1,6 @@
 import { PropsWithoutRef } from 'react'
 import { getStreams } from 'api/youtube/getStreams'
-import StreamListOfLive from 'features/hololive/stream/components/stream-list/StreamListOfLive'
+import StreamListOfScheduled from 'features/group/stream/components/stream-list/StreamListOfScheduled'
 
 type Props = {
   title: string
@@ -8,19 +8,22 @@ type Props = {
   compact?: boolean
 }
 
-export default async function Live({
+export default async function Schedule({
   title,
   description,
   compact
 }: PropsWithoutRef<Props>) {
   const streams = await getStreams({
-    status: 'live',
-    orderBy: [{ field: 'maxViewerCount', order: 'desc' }],
+    status: 'scheduled',
+    scehduledAfter: new Date(),
+    // +48 hours from now
+    scehduledBefore: new Date(new Date().getTime() + 48 * 60 * 60 * 1000),
+    orderBy: [{ field: 'scheduledStartTime', order: 'asc' }],
     limit: 100
   })
 
   return (
-    <StreamListOfLive
+    <StreamListOfScheduled
       streams={streams}
       title={title}
       description={description}
