+++ web/features/group/stream/components/badge/DurationBadge.tsx
